#29: Restructure project structure using the feature-based project structure without mvc design pattern.
