+++ src/main/java/org/weathertrack/validation/InputExceptionMessage.java
@@ -1,4 +1,4 @@
-package org.weathertrack.validation.model;
+package org.weathertrack.validation;
 
 public class InputExceptionMessage {
 	public static final String USER_INPUT_IS_NULL =
