+++ src/test/java/org/weathertrack/weather/WeatherDataTests.java
@@ -1,10 +1,9 @@
-package org.weathertrack.weather.model;
+package org.weathertrack.weather;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.weathertrack.ui.model.WeatherExceptionMessage;
 
 import java.util.stream.Stream;
