+++ src/test/java/org/weathertrack/weather/WeatherDisplayResourceTests.java
@@ -1,4 +1,4 @@
-package org.weathertrack.ui.model;
+package org.weathertrack.weather;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
